requirements.txt
Flask==3.0.0
flask-cors==4.0.0
deepface==0.0.92
pillow==10.1.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
tensorflow==2.15.0
gunicorn==21.2.0
